notebooks/DL0120EN-tensorflow/Python-Flask-MNIST-sample-app/app/requirements.txt
Flask==1.1.1
watson-machine-learning-client==1.0.378
Pillow==7.0.0
